QUIPU-664 fixes after review
