+++ src/modules/liquidity/pages/v3-item-page/components/position-details-create/index.ts
@@ -1,1 +1,1 @@
-export * from './position-details-create';
+export * from './pool-details-create';
