+++ src/modules/liquidity/pages/v3-item-page/components/position-details-create/pool-details-create.module.scss
