+++ src/modules/liquidity/pages/v3-item-page/components/position-details/index.ts
@@ -1,1 +1,1 @@
-export * from './position-details';
+export * from './pool-details';
