+++ src/modules/liquidity/pages/v3-item-page/components/position-details/pool-details.module.scss
