+++ src/modules/liquidity/pages/v3-item-page/create-new-position.page.tsx
@@ -1,7 +1,7 @@
 import { StickyBlock } from '@shared/components';
 import { useTranslation } from '@translation';
 
-import { PageTitleContainer, PositionDetailsCreate } from './components';
+import { PageTitleContainer, PoolDetailsCreate } from './components';
 
 export const CreateNewPositionPage = () => {
   const { t } = useTranslation();
@@ -10,7 +10,7 @@ export const CreateNewPositionPage = () => {
     <>
       <PageTitleContainer dataTestId="v3LiqCreatePosition" titleText={t('liquidity|createPosition')} />
       <StickyBlock>
-        <PositionDetailsCreate />
+        <PoolDetailsCreate />
       </StickyBlock>
     </>
   );
